[JAVA-12051] Moved code to core-java-streams-simple for ebook (#14889)
