+++ core-java-modules/core-java-11-2/README.md
@@ -5,7 +5,6 @@ This module contains articles about Java 11 core features
 ### Relevant articles
 - [Guide To Java 8 Optional](https://www.baeldung.com/java-optional)
 - [Guide to Java Reflection](http://www.baeldung.com/java-reflection)
-- [Guide to Java 8’s Collectors](https://www.baeldung.com/java-8-collectors)
 - [New Features in Java 11](https://www.baeldung.com/java-11-new-features)
 - [Getting the Java Version at Runtime](https://www.baeldung.com/get-java-version-runtime)
 - [Invoking a SOAP Web Service in Java](https://www.baeldung.com/java-soap-web-service)
