+++ core-java-modules/core-java-16/README.md
@@ -4,5 +4,4 @@
 - [Guide to mapMulti in Stream API](https://www.baeldung.com/java-mapmulti)
 - [Collecting Stream Elements into a List in Java](https://www.baeldung.com/java-stream-to-list-collecting)
 - [New Features in Java 16](https://www.baeldung.com/java-16-new-features)
-- [Guide to Java 8 groupingBy Collector](https://www.baeldung.com/java-groupingby-collector)
 - [Value-Based Classes in Java](https://www.baeldung.com/java-value-based-classes)
