--- core-java-modules/core-java-16/src/main/java/com/baeldung/java_16_features/groupingby/BlogPostType.java
@@ -1,5 +1,0 @@
-package com.baeldung.java_16_features.groupingby;
-
-public enum BlogPostType {
-    NEWS, REVIEW, GUIDE
-}
