+++ core-java-modules/core-java-streams-2/README.md
@@ -4,12 +4,9 @@ This module contains articles about the Stream API in Java.
 
 ### Relevant Articles:
 - [The Java 8 Stream API Tutorial](https://www.baeldung.com/java-8-streams)
-- [Introduction to Java 8 Streams](https://www.baeldung.com/java-8-streams-introduction)
 - [Java 8 Stream findFirst() vs. findAny()](https://www.baeldung.com/java-stream-findfirst-vs-findany)
-- [Guide to Stream.reduce()](https://www.baeldung.com/java-stream-reduce)
 - [Java IntStream Conversions](https://www.baeldung.com/java-intstream-convert)
 - [Java 8 Streams peek() API](https://www.baeldung.com/java-streams-peek-api)
-- [Working With Maps Using Streams](https://www.baeldung.com/java-maps-streams)
 - [Collect a Java Stream to an Immutable Collection](https://www.baeldung.com/java-stream-immutable-collection)
 - [How to Add a Single Element to a Stream](https://www.baeldung.com/java-stream-append-prepend)
 - [Operating on and Removing an Item from Stream](https://www.baeldung.com/java-use-remove-item-stream)
