--- core-java-modules/core-java-streams-2/src/main/java/com/baeldung/reduce/entities/User.java
@@ -1,25 +1,0 @@
-package com.baeldung.reduce.entities;
-
-public class User {
-
-    private final String name;
-    private final int age;
-
-    public User(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-    
-    @Override
-    public String toString() {
-        return "User{" + "name=" + name + ", age=" + age + '}';
-    }
-}
