+++ core-java-modules/core-java-streams-3/README.md
@@ -2,8 +2,7 @@
 
 This module contains articles about the Stream API in Java.
 
-### Relevant Articles: 
-- [The Difference Between map() and flatMap()](https://www.baeldung.com/java-difference-map-and-flatmap)
+### Relevant Articles:
 - [How to Use if/else Logic in Java 8 Streams](https://www.baeldung.com/java-8-streams-if-else-logic)
 - [The Difference Between Collection.stream().forEach() and Collection.forEach()](https://www.baeldung.com/java-collection-stream-foreach)
 - [Primitive Type Streams in Java 8](https://www.baeldung.com/java-8-primitive-streams)
@@ -12,5 +11,4 @@ This module contains articles about the Stream API in Java.
 - [Should We Close a Java Stream?](https://www.baeldung.com/java-stream-close)
 - [Returning Stream vs. Collection](https://www.baeldung.com/java-return-stream-collection)
 - [Convert a Java Enumeration Into a Stream](https://www.baeldung.com/java-enumeration-to-stream)
-- [When to Use a Parallel Stream in Java](https://www.baeldung.com/java-when-to-use-parallel-stream)
 - More articles: [[<-- prev>]](/../core-java-streams-2)
