+++ core-java-modules/core-java-streams-simple/README.md
@@ -1,0 +1,14 @@
+## Java Streams Ebook
+
+This module contains articles about Streams that are part of the Java Streams Ebook.
+
+### Relevant Articles
+
+- [Introduction to Java 8 Streams](https://www.baeldung.com/java-8-streams-introduction)
+- [Guide to Java 8’s Collectors](https://www.baeldung.com/java-8-collectors)
+- [Java Stream Filter with Lambda Expression](https://www.baeldung.com/java-stream-filter-lambda)
+- [Working With Maps Using Streams](https://www.baeldung.com/java-maps-streams)
+- [The Difference Between map() and flatMap()](https://www.baeldung.com/java-difference-map-and-flatmap)
+- [When to Use a Parallel Stream in Java](https://www.baeldung.com/java-when-to-use-parallel-stream)
+- [Guide to Java 8 groupingBy Collector](https://www.baeldung.com/java-groupingby-collector)
+- [Guide to Stream.reduce()](https://www.baeldung.com/java-stream-reduce)
