+++ core-java-modules/core-java-streams-simple/src/main/java/com/baeldung/streams/groupingby/BlogPostType.java
@@ -1,0 +1,5 @@
+package com.baeldung.streams.groupingby;
+
+public enum BlogPostType {
+    NEWS, REVIEW, GUIDE
+}
