+++ core-java-modules/core-java-streams-simple/src/main/java/com/baeldung/streams/reduce/entities/Review.java
@@ -1,0 +1,28 @@
+package com.baeldung.streams.reduce.entities;
+
+public class Review {
+
+    int points;
+    String review;
+
+    public Review(int points, String review) {
+        this.points = points;
+        this.review = review;
+    }
+
+    public int getPoints() {
+        return points;
+    }
+
+    public void setPoints(int points) {
+        this.points = points;
+    }
+
+    public String getReview() {
+        return review;
+    }
+
+    public void setReview(String review) {
+        this.review = review;
+    }
+}
