+++ core-java-modules/core-java-streams-simple/src/main/java/com/baeldung/streams/reduce/entities/User.java
@@ -1,0 +1,25 @@
+package com.baeldung.streams.reduce.entities;
+
+public class User {
+
+    private final String name;
+    private final int age;
+
+    public User(String name, int age) {
+        this.name = name;
+        this.age = age;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getAge() {
+        return age;
+    }
+    
+    @Override
+    public String toString() {
+        return "User{" + "name=" + name + ", age=" + age + '}';
+    }
+}
