+++ core-java-modules/core-java-streams-simple/src/test/java/com/baeldung/streams/Detail.java
