+++ core-java-modules/core-java-streams/README.md
@@ -9,7 +9,6 @@ This module contains articles about the Stream API in Java.
 - [Iterable to Stream in Java](https://www.baeldung.com/java-iterable-to-stream)
 - [How to Iterate Over a Stream With Indices](https://www.baeldung.com/java-stream-indices)
 - [Stream Ordering in Java](https://www.baeldung.com/java-stream-ordering)
-- [Java Stream Filter with Lambda Expression](https://www.baeldung.com/java-stream-filter-lambda)
 - [Counting Matches on a Stream Filter](https://www.baeldung.com/java-stream-filter-count)
 - [Summing Numbers with Java Streams](https://www.baeldung.com/java-stream-sum)
 - [How to Find All Getters Returning Null](https://www.baeldung.com/java-getters-returning-null)
